# docs/swagger.yaml
basePath: /api
definitions:
  audithandler.AuditLogResponse:
    properties:
      action:
        type: string
      actor:
        type: string
      details:
        additionalProperties: {}
        type: object
      id:
        type: integer
      resource:
        type: string
      timestamp:
        type: integer
    type: object
...
  databasehandler.DatabaseCreatePayload:
    properties:
      config:
        $ref: '#/definitions/databasehandler.ConfigPayload'
      content_type:
        type: string
      custom_fields:
        items:
          $ref: '#/definitions/databasehandler.DatabaseCustomField'
        type: array
      housekeeping:
        $ref: '#/definitions/databasehandler.HousekeepingPayload'
      name:
        type: string
    type: object
...
  databasehandler.DatabaseUpdatePayload:
    properties:
      config:
        $ref: '#/definitions/databasehandler.ConfigPayload'
      housekeeping:
        $ref: '#/definitions/databasehandler.HousekeepingPayload'
    type: object
...
  entryhandler.BulkDeleteRequest:
    properties:
      ids:
        items:
          type: integer
        type: array
    type: object
...
  tokenhandler.OidcTokenRequest:
    properties:
      idp_token:
        type: string
    type: object
  tokenhandler.TokenRequest:
    properties:
      refresh_token:
        type: string
    type: object
  tokenhandler.TokenResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
...
  utils.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  utils.MessageResponse:
    properties:
      message:
        type: string
    type: object
info:
  contact:
    email: denglerchr@gmail.com
    name: Christian Dengler
    url: https://www.swcd.lu
  description: This is a server for a image, audio and file storage with integrated
    web-ui.
  title: SWCD MediaHub-API
  version: 2.0.0-beta1
paths:
  /api/logout:
    post:
      consumes:
      - application/json
      description: Revokes a Refresh Token, effectively logging the user out.
      parameters:
      - description: Refresh token JSON body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/tokenhandler.TokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Logged out successfully message
          schema:
            $ref: '#/definitions/utils.MessageResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Logout
      tags:
      - token
  /api/token:
    post:
      consumes:
      - application/json
      description: |-
        Obtains an internal JWT Access/Refresh token pair.
        Supports two authentication methods:
        1. Local Authentication: Send standard Basic Auth headers.
        2. OIDC Token Exchange (commercial version only): Send a JSON body containing a valid external JWT (`idp_token`).
        Providing both methods in a single request will result in a 400 Bad Request.
      parameters:
      - description: OIDC Identity Provider Token (required if not using Basic Auth)
        in: body
        name: body
        schema:
          $ref: '#/definitions/tokenhandler.OidcTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Returns access and refresh tokens
          schema:
            $ref: '#/definitions/tokenhandler.TokenResponse'
        "400":
          description: Ambiguous authentication request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "401":
          description: Invalid credentials, invalid OIDC token, or missing authentication
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal server error or OIDC not available
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BasicAuth: []
      summary: Get a token pair
      tags:
      - token
  /api/token/refresh:
    post:
      consumes:
      - application/json
      description: Uses a valid Refresh Token to obtain a new Access/Refresh token
        pair.
      parameters:
      - description: Refresh token JSON body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/tokenhandler.TokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Returns new access and refresh tokens
          schema:
            $ref: '#/definitions/tokenhandler.TokenResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "401":
          description: Invalid or expired refresh token
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Refresh token pair
      tags:
      - token
  /audit:
    get:
      description: Retrieves a paginated list of system audit logs. This endpoint
        provides basic filtering by timestamp and pagination.
      parameters:
      - description: Number of logs to return (default 30)
        in: query
        name: limit
        type: integer
      - description: Offset for pagination (default 0)
        in: query
        name: offset
        type: integer
      - description: Sort order ('asc' or 'desc', default 'desc')
        in: query
        name: order
        type: string
      - description: Start timestamp (Unix milliseconds)
        format: int64
        in: query
        name: tstart
        type: integer
      - description: End timestamp (Unix milliseconds)
        format: int64
        in: query
        name: tend
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Returns an array of audit log objects
          schema:
            items:
              $ref: '#/definitions/audithandler.AuditLogResponse'
            type: array
        "400":
          description: Invalid parameter formats
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "403":
          description: Forbidden (Requires IsAdmin role)
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Failed to retrieve audit logs
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Get audit logs
      tags:
      - audit
  /database:
    post:
      consumes:
      - application/json
      description: Creates a new database with custom fields and a dedicated entry
        table.
      parameters:
      - description: Database Metadata
        in: body
        name: database
        required: true
        schema:
          $ref: '#/definitions/databasehandler.DatabaseCreatePayload'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/databasehandler.DatabaseResponse'
        "400":
          description: Invalid request payload or missing name
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "409":
          description: Database name already in use
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Failed to create database or storage folder
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BasicAuth: []
      summary: Create a new database
      tags:
      - database
  /database/{dbname}:
    delete:
      description: Deletes a database, its entry table, and all of its associated
        entries and metadata.
      parameters:
      - description: Database Name
        in: path
        name: dbname
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success message
          schema:
            $ref: '#/definitions/utils.MessageResponse'
        "400":
          description: Missing dbname path parameter or invalid name
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Database not found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Failed to delete database record or folder
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BasicAuth: []
      summary: Delete a database
      tags:
      - database
    get:
      description: Retrieves details, custom fields, and statistics for a specific
        database.
      parameters:
      - description: Database Name
        in: path
        name: dbname
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/databasehandler.DatabaseResponse'
        "400":
          description: Missing dbname path parameter
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Database not found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Failed to calculate stats
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BasicAuth: []
      summary: Get database details
      tags:
      - database
    put:
      consumes:
      - application/json
      description: Updates the housekeeping rules for a specific database.
      parameters:
      - description: Database Name
        in: path
        name: dbname
        required: true
        type: string
      - description: Housekeeping Rules or Config flag
        in: body
        name: housekeeping
        required: true
        schema:
          $ref: '#/definitions/databasehandler.DatabaseUpdatePayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/databasehandler.DatabaseResponse'
        "400":
          description: Invalid request payload or missing dbname path parameter
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Database not found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Failed to update database
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BasicAuth: []
      summary: Update database housekeeping rules
      tags:
      - database
  /database/{dbname}/entries:
    get:
      description: Retrieves a paginated list of entries from a specific database.
        Only supports time-based filters.
      parameters:
      - description: Database Name
        in: path
        name: dbname
        required: true
        type: string
      - description: Number of entries to return (default 30)
        in: query
        name: limit
        type: integer
      - description: Offset for pagination (default 0)
        in: query
        name: offset
        type: integer
      - description: Sort order ('asc' or 'desc', default 'desc')
        in: query
        name: order
        type: string
      - description: Start timestamp (Unix milliseconds)
        format: int64
        in: query
        name: tstart
        type: integer
      - description: End timestamp (Unix milliseconds)
        format: int64
        in: query
        name: tend
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Returns an array of entry metadata objects
          schema:
            items:
              $ref: '#/definitions/entryhandler.EntryResponse'
            type: array
        "400":
          description: Missing name param or invalid parameter formats
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "403":
          description: Forbidden (Requires CanView role)
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Database not found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Failed to retrieve entries
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BasicAuth: []
      summary: Get entries from a database (basic)
      tags:
      - database
  /database/{dbname}/entries/delete:
    post:
      consumes:
      - application/json
      description: Deletes multiple entries in a single atomic transaction. Updates
        database statistics only once.
      parameters:
      - description: Database Name
        in: path
        name: dbname
        required: true
        type: string
      - description: JSON object containing a list of Entry IDs to delete
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entryhandler.BulkDeleteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Summary of the deletion operation
          schema:
            $ref: '#/definitions/entryhandler.BulkDeleteResponse'
        "400":
          description: Invalid request, missing name, or empty IDs list
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "403":
          description: Forbidden (Requires CanDelete role)
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Database not found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Transaction failed
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BasicAuth: []
      summary: Bulk delete entries
      tags:
      - database
  /database/{dbname}/entries/export:
    post:
      consumes:
      - application/json
      description: Streams a ZIP archive containing the files and metadata (CSV) for
        the specified entries using io.Pipe.
      parameters:
      - description: Database Name
        in: path
        name: dbname
        required: true
        type: string
      - description: List of Entry IDs to export
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entryhandler.ExportRequest'
      produces:
      - application/zip
      responses:
        "200":
          description: ZIP Archive containing files and entries.csv
          schema:
            type: file
        "400":
          description: Missing name query parameter or empty IDs list
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "403":
          description: Forbidden (Requires CanView role)
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Database not found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: ZIP streaming failed
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BasicAuth: []
      summary: Export entries as ZIP
      tags:
      - database
  /database/{dbname}/entries/search:
    post:
      consumes:
      - application/json
      description: Retrieves a list of entry metadata matching the complex, nested
        filter criteria provided in the request body.
      parameters:
      - description: Database Name
        in: path
        name: dbname
        required: true
        type: string
      - description: JSON body defining filter, sort, and pagination logic
        in: body
        name: search
        required: true
        schema:
          $ref: '#/definitions/repository.SearchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Returns an array of matching results (even if empty)
          schema:
            items:
              $ref: '#/definitions/entryhandler.EntryResponse'
            type: array
        "400":
          description: Missing name, invalid JSON, missing limit, or invalid filter/sort
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "403":
          description: Forbidden (Requires CanView role)
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Database not found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BasicAuth: []
      summary: Search for entries in a database (complex)
      tags:
      - database
  /database/{dbname}/entry:
    post:
      consumes:
      - multipart/form-data
      description: |-
        Uploads a new entry to a specified database using multipart/form-data. The metadata part should be a JSON object containing the entry's timestamp, and any custom fields.
        The 'file' part's 'filename' in the Content-Disposition header will be extracted and saved.

        This endpoint uses a hybrid model:
        - **Small files (<= Configured Limit):** Processed synchronously. Returns `201 Created` with the full entry metadata.
        - **Large files (> Configured Limit):** Processed asynchronously. Returns `202 Accepted` with a partial response. The client should poll `GET /api/entry/meta` until the `status` field is 'ready'.
      parameters:
      - description: Database Name
        in: query
        name: database_name
        required: true
        type: string
      - description: JSON metadata for the entry
        in: formData
        name: metadata
        required: true
        type: string
      - description: Entry file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: For small files (synchronous processing)
          schema:
            $ref: '#/definitions/entryhandler.EntryResponse'
        "202":
          description: For large files (asynchronous processing)
          schema:
            $ref: '#/definitions/entryhandler.PartialEntryResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Database not found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "415":
          description: Unsupported entry format
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BasicAuth: []
      summary: Upload an entry
      tags:
      - entry
  /database/{dbname}/entry/{id}:
    delete:
      description: Deletes an entry file from disk and its metadata from the database.
      parameters:
      - description: Database Name
        in: path
        name: dbname
        required: true
        type: string
      - description: Entry ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success message
          schema:
            $ref: '#/definitions/utils.MessageResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Database or entry not found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BasicAuth: []
      summary: Delete an entry
      tags:
      - entry
    get:
      description: Retrieves all metadata for a single entry, including custom fields.
      parameters:
      - description: Database Name
        in: path
        name: dbname
        required: true
        type: string
      - description: Entry ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: The full entry metadata object
          schema:
            $ref: '#/definitions/entryhandler.EntryResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Database or entry not found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BasicAuth: []
      summary: Get entry metadata
      tags:
      - entry
    patch:
      consumes:
      - application/json
      description: Updates an entry's mutable metadata, including custom fields, the
        'timestamp' and the 'filename'.
      parameters:
      - description: Database Name
        in: path
        name: dbname
        required: true
        type: string
      - description: Entry ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      - description: JSON object with fields to update
        in: body
        name: updates
        required: true
        schema:
          $ref: '#/definitions/entryhandler.PostPatchEntryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: The full, updated entry metadata object
          schema:
            $ref: '#/definitions/entryhandler.EntryResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Database or entry not found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BasicAuth: []
      summary: Update entry metadata
      tags:
      - entry
  /database/{dbname}/entry/{id}/file:
    get:
      description: Retrieves a raw entry file. Supports Content Negotiation (JSON
        vs Binary) and HTTP Range Requests (Streaming).
      parameters:
      - description: Database Name
        in: path
        name: dbname
        required: true
        type: string
      - description: Entry ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      - description: Byte range request (e.g., bytes=0-1023)
        in: header
        name: Range
        type: string
      produces:
      - application/octet-stream
      - application/json
      responses:
        "200":
          description: 'Base64 encoded file data (if Accept: application/json)'
          headers:
            Accept-Ranges:
              description: bytes
              type: string
          schema:
            $ref: '#/definitions/entryhandler.FileJSONResponse'
        "206":
          description: Partial content (streaming response)
          headers:
            Accept-Ranges:
              description: bytes
              type: string
            Content-Range:
              description: bytes start-end/total
              type: string
          schema:
            type: file
        "400":
          description: Invalid request or ID format
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Database or entry not found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "409":
          description: File is currently processing
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "416":
          description: Range Not Satisfiable
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Get an entry file
      tags:
      - entry
  /database/{dbname}/entry/{id}/preview:
    get:
      description: Retrieves a 200x200 WebP preview of an entry. Supports Content
        Negotiation via Accept header.
      parameters:
      - description: Database Name
        in: path
        name: dbname
        required: true
        type: string
      - description: Entry ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - image/webp
      - application/json
      responses:
        "200":
          description: 'Base64 encoded preview data (if Accept: application/json)'
          schema:
            $ref: '#/definitions/entryhandler.FileJSONResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Database, entry, or preview not found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BasicAuth: []
      summary: Get an entry preview
      tags:
      - entry
...
  /databases:
    get:
      description: Retrieves a list of all available databases and their statistics.
      produces:
      - application/json
      responses:
        "200":
          description: Returns an empty array if no databases exist
          schema:
            items:
              $ref: '#/definitions/databasehandler.DatabaseResponse'
            type: array
        "500":
          description: Failed to retrieve databases
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BasicAuth: []
      summary: List all databases
      tags:
      - database
...
schemes:
- http
securityDefinitions:
  BasicAuth:
    type: basic
  BearerAuth:
    description: Type "Bearer" followed by a space and a JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"